created migrations to model the database
