+++ src/database/20241202140951141_knex_migrations.ts
@@ -1,0 +1,14 @@
+import { Knex } from "knex";
+
+export async function up(knex: Knex): Promise<void> {
+  await knex.schema.createTable("user", function (table) {
+    table.increments("id").primary();
+    table.string("name");
+    table.string("email").unique();
+    table.string("password");
+  });
+}
+
+export async function down(knex: Knex): Promise<void> {
+  await knex.schema.dropTableIfExists("user");
+}
